app/ingest-api/requirements.txt
fastapi
uvicorn[standard]
google-cloud-storage
python-dotenv
requests
faker
pydantic-settings